Replace an apostrophe; fix some line breaks
Specifically for the line breaks, Markdown ignores intra-paragraph line breaks, so the signature was showing as one long line in the rendered version.  Markdown supports standard line breaks by adding two trailing spaces to the end of the line, but this invisible signaling is easy to miss and/or accidentally erase, so I opted to use explicit `<br>`s instead (also supported by Markdown).